[math/py-pinocchio] Fix patch-ak for patch --posix
With --posix option, patch will patch the _old_ file if such a file is
present. Also, patch is documented to never make any backup files in --posix
mode, but experience shows that in case of patch fuzz, is still makes one.

Given this, it happended that patch-ag and patch-ak where patching the same
file, with patch-ag applying with some fuzz factor, thus making a backup file
when using --posix and leading to patch-ak applying to the backup file instead
of the good one (also because of --posix).

Anyway, it's not OK in robotpkg to have to patches patching the same file, so
merge the two patches in one, fixing the issue.

Thanks to Gialuca for reporting this and helping in solving it!