[mk/sysdep] Use ffmpeg version for ffmpeg libraries
On most systems, ffmpeg libav* libraries are part of ffmpeg and share the same
packaging version as ffmpeg. While the sysdep file for individual libraries
works nicely, it breaks the binary deb package production because of the
incompatible version requirements (e.g. libavdevice>=60 coming with ffmpeg-6
should be libavdevice>=6 on Debian).